Move constraints to their own file
