+++ lang/elaborator/src/unifier/mod.rs
@@ -1,2 +1,3 @@
 pub mod dec;
 pub mod unify;
+pub mod constraints;
